move function to separate file
