+++ tests/api_apple.spec.js
@@ -2,7 +2,7 @@ const chai = require('chai');
 const chaiHttp = require('chai-http');
 const app = require('../server');
 const should = chai.should();
-const { testBasicProperties } = require('./mochaSetup.spec');
+const { testBasicProperties } = require('./testingFunctions');
 
 chai.use(chaiHttp);
